RUD authorization Pokemon

+++ Routes/Pokemon.js
@@ -1,14 +1,20 @@
 const express = require("express");
-const authenticateToken = require("../TokenService/Auth")
+const authenticateToken = require("../TokenService/Auth");
 const PokemonModel = require("../Models/Pokemon");
 
 const router = express.Router();
 
-
-//Get All Pokemon
-router.get("/", authenticateToken.authenticateToken, async (req, res, next) => {
+//Get All Pokemon by Trainer
+router.get("/", authenticateToken.authenticateToken, async (req, res) => {
   try {
-    const pokemon = await PokemonModel.find();
+    const authHeader = req.headers["authorization"];
+    const token = authHeader && authHeader.split(" ")[1];
+    if (token == null) {
+      return res.status(401).json({ message: "authorization missing" });
+    }
+    var idTrainer = await authenticateToken.decodeToken(token);
+
+    const pokemon = await PokemonModel.find({ trainerId: idTrainer });
     res.status(200).json(pokemon);
   } catch (error) {
     res.status(500).json({ message: error.message });
@@ -26,7 +32,14 @@ router.get("/:id", authenticateToken.authenticateToken, async (req, res) => {
 });
 
 //Create Pokemon
-router.post("/", authenticateToken.authenticateToken, async (req, res, next) => {
+router.post("/", authenticateToken.authenticateToken, async (req, res) => {
+  const authHeader = req.headers["authorization"];
+  const token = authHeader && authHeader.split(" ")[1];
+  if (token == null) {
+    return res.status(401).json({ message: "authorization missing" });
+  }
+  var idTrainer = await authenticateToken.decodeToken(token);
+
   const pokemon = new PokemonModel({
     name: req.body.name,
     gender: req.body.gender,
@@ -40,6 +53,7 @@ router.post("/", authenticateToken.authenticateToken, async (req, res, next) => 
     type: {
       name: req.body.type.name,
     },
+    trainerId: idTrainer,
   });
   try {
     const pokemonSave = await pokemon.save();
@@ -52,31 +66,61 @@ router.post("/", authenticateToken.authenticateToken, async (req, res, next) => 
 //Update Pokemon
 router.patch("/:id", authenticateToken.authenticateToken, async (req, res) => {
   try {
+    const authHeader = req.headers["authorization"];
+    const token = authHeader && authHeader.split(" ")[1];
+    if (token == null) {
+      return res.status(401).json({ message: "authorization missing" });
+    }
+    var idTrainer = await authenticateToken.decodeToken(token);
+
     const id = req.params.id;
     const updatedPokemon = req.body;
     const options = { new: true };
 
-    const result = await PokemonModel.findByIdAndUpdate(
-      id,
-      updatedPokemon,
-      options
-    );
+    const pokemonById = await PokemonModel.findById(id);
 
-    res.send(result);
+    if (pokemonById.trainerId == idTrainer) {
+      const result = await PokemonModel.findByIdAndUpdate(
+        id,
+        updatedPokemon,
+        options
+      );
+      res.send(result);
+    } else {
+      return res.status(401).json({ message: "Not authorized >_<" });
+    }
   } catch (error) {
     res.status(400).json({ message: error.message });
   }
 });
 
 //Delete pokemon
-router.delete("/:id", authenticateToken.authenticateToken, async (req, res, next) => {
-  try {
-    const id = req.params.id;
-    const result = await PokemonModel.findByIdAndDelete(id);
-    res.status(200).json(result);
-  } catch (error) {
-    res.status(400).json({ message: error.message });
+router.delete(
+  "/:id",
+  authenticateToken.authenticateToken,
+  async (req, res) => {
+    try {
+      const authHeader = req.headers["authorization"];
+      const token = authHeader && authHeader.split(" ")[1];
+      if (token == null) {
+        return res.status(401).json({ message: "authorization missing" });
+      }
+      var idTrainer = await authenticateToken.decodeToken(token);
+
+      const id = req.params.id;
+
+      const pokemonById = await PokemonModel.findById(id);
+
+      if (pokemonById.trainerId == idTrainer) {
+        const result = await PokemonModel.findByIdAndDelete(id);
+        res.status(200).json({deleted:result});
+      } else {
+        return res.status(401).json({ message: "Not authorized >_<" });
+      }
+    } catch (error) {
+      res.status(400).json({ message: error.message });
+    }
   }
-});
+);
 
 module.exports = router;
